+++ internal/repp/frag.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/Lattice-Automation/repp/config"
+	"github.com/Lattice-Automation/repp/internal/config"
 	"github.com/jinzhu/copier"
 )
