+++ internal/repp/primer3.go
@@ -10,7 +10,7 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/Lattice-Automation/repp/config"
+	"github.com/Lattice-Automation/repp/internal/config"
 )
 
 // primer3 is a utility struct for executing primer3 to create primers on a fragment
